Develop Into Master (#17)
* Feature1 - partially ready

This is the first commit of the first feature.

* Finished working on repository 1.

This is the final commit in feature 1. I will now ask for a pull request
to the Develop branch.

* Feature 2

Second Feature. I write this at the same time a different user is
working on feature 3.

* Feature 3 implemented at the same time as feature2

* Last Commit in feature 3 before pull request

* Last Commit in Feature2 before pull request

* Feature 4 (after 1.0.0.1 was sent to QA)

* Fix for issue1. (#9)

Merge the fixes in 1.0.0.1 (in this case, the fix to issue 1) to Develop branch.

* Feature5 (#10)

Approved!

* Fixed the issue. (#12)

Approved.

* Add my mirror clone directory to ignore...

* Fixes #14

Approved.

* Create TestFile

* Create local file

* Update TestFile

* Changeee